Add clear method to bitwise variables
